move function and book class into a module
