+++ index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="style.css" />
+    <link rel="stylesheet" href="index.css" />
     <title>Book List</title>
   </head>
   <body>
@@ -50,6 +50,6 @@
         </ul>
       </div>
     </section>
-    <script src="main.js"></script>
+    <script src="index.js" type="module"></script>
   </body>
 </html>
