+++ package.json
@@ -19,5 +19,13 @@
     "stylelint-config-standard": "^21.0.0",
     "stylelint-csstree-validator": "^1.9.0",
     "stylelint-scss": "^3.21.0"
-  }
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Sirvelasque/books-with-es6.git"
+  },
+  "bugs": {
+    "url": "https://github.com/Sirvelasque/books-with-es6/issues"
+  },
+  "homepage": "https://github.com/Sirvelasque/books-with-es6#readme"
 }
